requirements.txt
tensorflow==2.10.0
keras==2.10.0
numpy
opencv-python
pillow